+++ app/components/SignUp/CustomerInfo/CustomerInfoForm.js
@@ -1,37 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { change, FormSection, formValueSelector, reduxForm } from 'redux-form';
+import { FormSection, formValueSelector, reduxForm } from 'redux-form';
 import { withTranslation } from 'react-i18next';
 import { withRouter } from 'react-router';
 import Submit from '../../helix/buttons/Submit';
 import { validateCustomerInformation } from '../../../validators';
 import CustomerType from './CustomerType';
 import Product from './Product';
-import { ADDRESS_FIELDS } from '../../../actions/constants/address';
 
 export class CustomerInfoForm extends React.Component {
-  handleChange = (e) => {
-    if (e.target.value === 'rbu') {
-      this.populateAddressFields();
-      this.props.clearProduct();
-    } else {
-      this.clearAddressFields();
-    }
-  };
-
-  populateAddressFields = () => {
-    Object.entries(ADDRESS_FIELDS).forEach((entry) => {
-      this.props.setAddress(...entry);
-    });
-  };
-
-  clearAddressFields = () => {
-    Object.keys(ADDRESS_FIELDS).forEach((field) => {
-      this.props.setAddress(field, '');
-    });
-  };
-
   onSubmit = () => {
     this.props.history.push('/billing');
   };
@@ -65,24 +43,11 @@ export class CustomerInfoForm extends React.Component {
 
 CustomerInfoForm.propTypes = {
   t: PropTypes.func.isRequired,
-  clearProduct: PropTypes.func.isRequired,
-  setAddress: PropTypes.func.isRequired,
   customerType: PropTypes.string,
   handleSubmit: PropTypes.func.isRequired,
   history: PropTypes.shape({
     push: PropTypes.func.isRequired
   })
-};
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    clearProduct: () => {
-      dispatch(change('signUp', 'customerInfo.productType', ''));
-    },
-    setAddress: (field, value) => {
-      dispatch(change('signUp', `billingInfo.address.${field}`, value));
-    }
-  };
 };
 
 const mapStateToProps = (state) => {
@@ -105,4 +70,4 @@ const CustomerInformationReduxForm = reduxForm({
   forceUnregisterOnUnmount: true // <------ unregister fields on unmount
 })(withTranslation()(CustomerInfoForm));
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(CustomerInformationReduxForm));
+export default withRouter(connect(mapStateToProps)(CustomerInformationReduxForm));
